Retry token backgroundsync and background sync for draft delete (#44)
Retry token backgroundsync and background sync for draft delete (#44)

+++ forms-flow-rsbcservice/src/services/offlineSubmissions.ts
@@ -10,7 +10,7 @@ import {
   OfflineDeleteService,
   OfflineFetchService
 } from "../formsflow-rsbcservices";
-import { OfflineSubmission } from "../storage/ffDb";
+import { DeletedDraft, OfflineSubmission } from "../storage/ffDb";
 import {
   FormData,
   FormioCreateResponse,
@@ -26,7 +26,6 @@ class OfflineSubmissions {
     try {
       // Call token refresh.
       const tenantId = localStorage.getItem("tenantKey") ?? "";
-      console.log(tenantId, "_tenantId");
       let instance = KeycloakService.getInstance(
         KEYCLOAK_URL_AUTH,
         KEYCLOAK_URL_REALM,
@@ -47,12 +46,48 @@ class OfflineSubmissions {
       // Process drafts and submissions concurrently using Promise.all
       const processDraftsPromise = this.processDrafts(submissions);
       const processSubmissionPromise = this.processSubmission(submissions);
-
+      const deleteDraftPromise = this.processDraftDelete();
       // Wait for both processes to finish
-      await Promise.all([processDraftsPromise, processSubmissionPromise]);
+      await Promise.all([
+        processDraftsPromise,
+        processSubmissionPromise,
+        deleteDraftPromise
+      ]);
     } catch (error) {
       console.error("Error processing drafts or submissions:", error);
     }
+  }
+
+  /**
+   * Process the draft delete.
+   */
+  private static async processDraftDelete() {
+    // Fetch all drafts to be deleted from the local db
+    const draftsToBeDeleted = await OfflineFetchService.fetchAllDraftDelete();
+    const serverDraftIdToBeDeletedList = await this.getServerDraftIdList(
+      draftsToBeDeleted
+    );
+    await this.deleteDrafts(serverDraftIdToBeDeletedList);
+  }
+
+  /**
+   * Function to process delete drafts.
+   */
+  private static async deleteDrafts(serverDraftIdToBeDeletedList: number[]) {
+    for (const serverId of serverDraftIdToBeDeletedList) {
+      const URL = `${WEB_BASE_URL}/draft/${serverId}`;
+      // Calling the API
+      await RequestService.httpDELETERequest(URL);
+      // Delete the local indexdb entries.
+      await OfflineDeleteService.deleteDraftDeleteWithServerId(serverId);
+    }
+  }
+
+  /**
+   * Get server draft id list.
+   */
+  private static async getServerDraftIdList(draftsToBeDeleted: DeletedDraft[]) {
+    return draftsToBeDeleted.map((draft) => draft.serverDraftId);
   }
 
   /**
@@ -94,13 +129,18 @@ class OfflineSubmissions {
   private static async prepareAndSubmitDraft(
     draft: OfflineSubmission
   ): Promise<void> {
-    const url = `${WEB_BASE_URL}/draft`;
-    const payload = {
-      data: draft.data,
-      formId: draft.formId
-    };
-    await RequestService.httpPOSTRequest(url, payload);
-    await this.deleteLocalSubmissions(draft);
+    try {
+      const url = `${WEB_BASE_URL}/draft`;
+      const payload = {
+        data: draft.data,
+        formId: draft.formId
+      };
+      await RequestService.httpPOSTRequest(url, payload);
+      await this.deleteLocalSubmissions(draft);
+      throw new Error("Request failed");
+    } catch (error) {
+      console.error("Error creating the submission:", error);
+    }
   }
 
   /**
@@ -110,13 +150,18 @@ class OfflineSubmissions {
   private static async prepareAndUpdateDraft(
     draft: OfflineSubmission
   ): Promise<void> {
-    const url = `${WEB_BASE_URL}/draft/${draft.serverDraftId}`;
-    const payload = {
-      data: draft.data,
-      formId: draft.formId
-    };
-    await RequestService.httpPUTRequest(url, payload);
-    await this.deleteLocalSubmissions(draft);
+    try {
+      const url = `${WEB_BASE_URL}/draft/${draft.serverDraftId}`;
+      const payload = {
+        data: draft.data,
+        formId: draft.formId
+      };
+      await RequestService.httpPUTRequest(url, payload);
+      await this.deleteLocalSubmissions(draft);
+      throw new Error("Request failed");
+    } catch (error) {
+      console.error("Error creating the submission:", error);
+    }
   }
 
   /**
@@ -162,7 +207,8 @@ class OfflineSubmissions {
       const response: FormioCreateResponse =
         await this.prepareAndSubmitFormioSubmission(data);
       await this.triggerApplicationCreate(response);
-      await this.deleteLocalSubmissions(data);
+      await this.deleteLocalSubmissions(data, true);
+      throw new Error("Request failed");
     } catch (error) {
       console.error("Error creating the submission:", error);
     }
@@ -175,25 +221,21 @@ class OfflineSubmissions {
   private static async prepareAndSubmitFormioSubmission(
     data: OfflineSubmission
   ): Promise<any> {
-    try {
-      const formioUrl = `${API_URL}/form/${data.formId}/submission`;
-      const formioPayload = {
-        data: data.data,
-        metadata: data.submissionData?.metadata,
-        state: data.submissionData?.state,
-        _vnote: data.submissionData?._vnote
-      };
-      const header = { "x-jwt-token": localStorage.getItem("formioToken") };
-      return RequestService.httpPOSTRequest(
-        formioUrl,
-        formioPayload,
-        null,
-        false,
-        header
-      );
-    } catch (error) {
-      console.error("Error creating the formio submission:", error);
-    }
+    const formioUrl = `${API_URL}/form/${data.formId}/submission`;
+    const formioPayload = {
+      data: data.data,
+      metadata: data.submissionData?.metadata,
+      state: data.submissionData?.state,
+      _vnote: data.submissionData?._vnote
+    };
+    const header = { "x-jwt-token": localStorage.getItem("formioToken") };
+    return RequestService.httpPOSTRequest(
+      formioUrl,
+      formioPayload,
+      null,
+      false,
+      header
+    );
   }
 
   /**
@@ -203,16 +245,12 @@ class OfflineSubmissions {
   private static async triggerApplicationCreate(
     data: FormioCreateResponse
   ): Promise<void> {
-    try {
-      // Process and transform the data.
-      const submissionData = this.prepareApplicationPayload(data);
-      // API URL
-      const URL = `${WEB_BASE_URL}/application/create`;
-      // Calling the API
-      await RequestService.httpPOSTRequest(URL, submissionData);
-    } catch (error) {
-      console.error("Error creating the submission:", error);
-    }
+    // Process and transform the data.
+    const submissionData = this.prepareApplicationPayload(data);
+    // API URL
+    const URL = `${WEB_BASE_URL}/application/create`;
+    // Calling the API
+    await RequestService.httpPOSTRequest(URL, submissionData);
   }
 
   /**
@@ -244,7 +282,8 @@ class OfflineSubmissions {
       const response: FormioCreateResponse =
         await this.prepareAndSubmitFormioSubmission(data);
       await this.triggerApplicationUpdate(response, data?.serverDraftId);
-      await this.deleteLocalSubmissions(data);
+      await this.deleteLocalSubmissions(data, true);
+      throw new Error("Request failed");
     } catch (error) {
       console.error("Error creating and updating the submission:", error);
     }
@@ -259,16 +298,12 @@ class OfflineSubmissions {
     data: FormioCreateResponse,
     draft_id: number | string
   ): Promise<void> {
-    try {
-      // Process and transform the data.
-      const submissionData = this.prepareApplicationPayload(data);
-      // API URL
-      const URL = `${WEB_BASE_URL}/draft/${draft_id}/submit`;
-      // Calling the API
-      await RequestService.httpPUTRequest(URL, submissionData);
-    } catch (error) {
-      console.error("Error creating the submission:", error);
-    }
+    // Process and transform the data.
+    const submissionData = this.prepareApplicationPayload(data);
+    // API URL
+    const URL = `${WEB_BASE_URL}/draft/${draft_id}/submit`;
+    // Calling the API
+    await RequestService.httpPUTRequest(URL, submissionData);
   }
 
   /**
@@ -276,12 +311,15 @@ class OfflineSubmissions {
    * @param submissions offline submissions that need to be deleted after the process.
    */
   private static async deleteLocalSubmissions(
-    submission: OfflineSubmission
+    submission: OfflineSubmission,
+    deleteApplications = false
   ): Promise<void> {
     await OfflineDeleteService.deleteOfflineSubmission(submission._id);
-    await OfflineDeleteService.deleteApplicationWithLocalSubmissionId(
-      submission.localSubmissionId
-    );
+    if (deleteApplications) {
+      await OfflineDeleteService.deleteApplicationWithLocalSubmissionId(
+        submission.localSubmissionId
+      );
+    }
   }
 
   /**
